feat(db): store handicap as numeric(3,1)
Change players.handicap from integer to numeric(3,1) so it can hold a
one-decimal handicap index (e.g. 21.3) sourced from TheGrint, instead of
a whole number. Drop the .int() constraint from the player create/update
validation so decimal handicaps are accepted in the admin UI.

Includes the generated 0002 migration (integer values cast to numeric).

+++ src/lib/server/players.ts
@@ -12,7 +12,7 @@ const PIN_PATTERN = /^\d{4}$/;
 
 const createPlayerSchema = z.object({
   name: z.string().trim().min(1, "Name is required"),
-  handicap: z.coerce.number().int().min(0).default(0),
+  handicap: z.coerce.number().min(0).default(0),
   pin: z.string().regex(PIN_PATTERN, "PIN must be 4 digits"),
   photoUrl: z.string().nullable().optional(),
 });
@@ -20,7 +20,7 @@ const createPlayerSchema = z.object({
 const updatePlayerSchema = z
   .object({
     name: z.string().trim().min(1).optional(),
-    handicap: z.coerce.number().int().min(0).optional(),
+    handicap: z.coerce.number().min(0).optional(),
     pin: z.string().regex(PIN_PATTERN).optional(),
     photoUrl: z.string().nullable().optional(),
   })
